patch(2.0): updating the website into the latest values

+++ app/Livewire/Admin/DocumentType/DocumentTypeList.php
@@ -24,8 +24,7 @@ class DocumentTypeList extends Component
     public $count = 0;
 
     public $file;
-
-    public $lastOrder;
+ 
 
 
     protected $listeners = [
@@ -35,9 +34,19 @@ class DocumentTypeList extends Component
     ];
 
 
-    public function mount(){ 
+    public function mount(){
+        
+        
+
+        // dd($this->lastOrder);
+
+        // $this->resetOrder();
     }
 
+
+    public function getLastOrderProperty(){
+        return DocumentType::max('order') ?? 0;
+    }
  
 
     // Method to delete selected records
@@ -82,7 +91,18 @@ class DocumentTypeList extends Component
         $document_type = DocumentType::find($id);
 
 
+        // dd($document_type->reviewers()->exists());
+  
+        // Check if document type has related records
+        if ($document_type->project_documents()->exists() || $document_type->reviewers()->exists()) {
+            Alert::error('Error', 'Cannot delete document type because it has related records such as projects and reviewers. ');
+            return redirect()->route('document_type.index');
+        }
+
+
+
         $document_type->delete();
+        $this->resetOrder();
  
         // ActivityLog::create([
         //     'log_action' => "Global Project reviewer '".$document_type->name."' on list deleted ",
@@ -119,6 +139,12 @@ class DocumentTypeList extends Component
             case "Name Z - A":
                 $query = $query->orderBy('name', 'DESC');
                 break;
+            case "Order 1 - 100":
+                $query = $query->orderBy('order', 'ASC');
+                break;
+            case "Order 100 - 1":
+                $query = $query->orderBy('order', 'DESC');
+                break;
             case "Latest Added":
                 $query = $query->orderBy('document_types.created_at', 'DESC');
                 break;
@@ -132,22 +158,87 @@ class DocumentTypeList extends Component
                 $query = $query->orderBy('document_types.updated_at', 'ASC');
                 break;
             default:
-                $query = $query->orderBy('document_types.updated_at', 'DESC');
+                $query = $query->orderBy('document_types.order', 'ASC');
         }
 
         return $query->paginate($this->record_count);
     }
+
+
+    public function updateOrder($reviewer_id, $order, $direction){
+
+        if($direction == "move_up"){
+
+
+            $new_order = $order - 1; // minus because the direction is 1 to 100
+
+            //Updated
+            $prev_reviewer = DocumentType::where('order', $new_order)->first();
+            $prev_reviewer->order = $order;
+            $prev_reviewer->save();
+
+            /**update the value  */
+
+            $reviewer = DocumentType::find($reviewer_id);
+            $reviewer->order = $new_order;
+            $reviewer->save();
+
+
+
+
+
+        }else if($direction == "move_down"){
+
+
+            $new_order = $order + 1;  // add because the direction is 1 to 100
+
+            //Updated
+            $prev_reviewer = DocumentType::where('order', $new_order)->first();
+            $prev_reviewer->order = $order;
+            $prev_reviewer->save();
+
+            /**update the value  */
+
+            $reviewer = DocumentType::find($reviewer_id);
+            $reviewer->order = $new_order;
+            $reviewer->save();
+
+
+
+        }
+
+
+        $this->resetOrder();
+
+
+    }
+
+
+    public function resetOrder()
+    {
+        $document_types = DocumentType::orderBy('order', 'ASC')
+            ->get();
+    
+        foreach ($document_types as $index => $document_type) {
+            $document_type->order = $index + 1;
+            $document_type->save();
+        }
+    }
+
+
+
 
     public function render()
     {
 
          
 
- 
+        
 
 
         return view('livewire.admin.document-type.document-type-list',[
-            'document_types' => $this->document_types 
+            'document_types' => $this->document_types ,
+            'lastOrder' => $this->lastOrder ,
         ]);
     }
     // public function render()
